functions/train_model/requirements.txt
google-api-python-client
oauth2client
google-cloud-error-reporting
flask